Module elastic_responses::parsing

Response type parsing.

Structs

Buffered
A response body that has been buffered.

HttpResponseHead
The non-body component of the HTTP response.

MaybeOkResponse
A response that might be successful or an ApiError.

Parse
A parser that separates taking a response type from the readable body type.

Unbuffered
A response body that hasn't been buffered yet.

Enums

MaybeBufferedResponse
A response body that may or may not have been buffered.

Traits

IsOk
Convert a response message into a either a success or failure result.

ResponseBody
A http response body that can be buffered into a json value.

Functions

parse