Enum elastic_responses::parsing::MaybeBufferedResponse

pub enum MaybeBufferedResponse<B> where
    B: ResponseBody,  {
    Unbuffered(B),
    Buffered(B::Buffered),
    Value(Value),
}

A response body that may or may not have been buffered.

This type makes it possible to inspect the response body for an error type before passing it along to be deserialised properly.

Variants
Unbuffered(B)Buffered(B::Buffered)Value(Value)