Struct elastic_responses::parsing::MaybeOkResponse

pub struct MaybeOkResponse<B> where
    B: ResponseBody,  { /* fields omitted */ }

A response that might be successful or an ApiError.

Methods

impl<B> MaybeOkResponse<B> where
B: ResponseBody,

pub fn new<I>(ok: bool, res: I) -> Self where
I: Into<MaybeBufferedResponse<B>>,

Create a new response that indicates where or not the body is successful or an ApiError.

pub fn ok<I>(res: I) -> Self where
I: Into<MaybeBufferedResponse<B>>,

Create a response where the body is successful.

pub fn err<I>(res: I) -> Self where
I: Into<MaybeBufferedResponse<B>>,

Create a resposne where the body is an error.